settings/requirements.txt
Django==4.2
python-dateutil==2.8.2
pydicom==2.3.1
Pillow==9.5.0
numpy==1.24.3
reportlab==3.6.12
psycopg2==2.9.6
djangorestframework==3.14.0